fixed reiew comments
