+++ webui/src/pages/repositories/repository/fileRenderers/types.tsx
@@ -1,0 +1,27 @@
+export interface RendererComponent {
+    repoId: string;
+    refId: string;
+    path: string;
+    fileExtension: string | null;
+    contentType: string | null;
+    sizeBytes: number;
+}
+
+export interface RendererComponentWithText extends RendererComponent{
+    text: string;
+}
+
+export interface RendererComponentWithTextCallback extends RendererComponent {
+    onReady: (text: string) => JSX.Element;
+}
+
+export enum FileType {
+    DATA,
+    MARKDOWN,
+    IPYNB,
+    IMAGE,
+    PDF,
+    TEXT,
+    UNSUPPORTED,
+    TOO_LARGE,
+}
